+++ CHANGELOG.md
@@ -12,7 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Tags tree
+- Filters: year, month and day
+
+### Changed
+
+- Swing instead of AWT
 
 ## [0.2.0] - 2024-01-31